Implement and use the syscall trampoline.
Some syscalls are trapped by the filter, however, we need to redo it
in the signal handler of SIGSYS.  If we call syscalls again in the
signal handler, it will be trapped again, and never success.  By
setting a trampoline code at a fixed address range (0x200000000000 ~
+4096), the filter always skip/allow all syscalls makde from this
range.  The signal handler call this trampoline to make syscalls to by
pass the filter.