Don't use new and delete operators to allocate or free memory.
GCC has builtin implementations for these operators.  They can be
overwritten, but its declaration and signature made them visible out
of the shared object.  It make GOT happend, and we don't want it.  I
have no idea how to avoid it.  To keep it simple, I just avoid to use
it to allocate memory.