Handle null terminated list of init functions in the loader.
This patch change the format of the list of init functions.  The
loader doesn't receive the number of init functions, instead the list
is null terminated.

The shell code shellcode_funcall_trap that call the loader caused
segfualt sometime.  The reason of segfaul is unknow yet, but if the
bytes before an instruction are not mapped segfault may occur, however
not everytime.  The solution is to add 2 nop instructions at begining,
and jump to the instruction after them directly.