Use global_flags to switch features.
By hacking relocation, the value of global_flags can be changed from
the flightdeck.  However, the real address of the variable will be
added to it's value, that is what relactions are supposed to be.

To neutralize the effects of relocations, it should subtract the value
of the variable with it's address to recover it's real value.