Fix the case that execve() fails.
For the case that execve() fails, it doesn't cause SIGSTRAP.  So, the
Command Center doesn't receive any events and wait forever.

This patch add a handler for execve() to make a SIGTRAP by calling
kill() to notify the Command Center.  By checking the event of the
signal; 0, not PTRACE_EVENT_EXEC, the Command Center know that
execve() fails.